exercicio2.py

+++ Conta_Bancaria/exercicios_teste/exercicio2.py
@@ -3,9 +3,9 @@
 Faça um algoritmo para ler três números e imprimir a soma, média e produto dos números lidos.
 
 '''
-valor1 = int(input('Digite o primeiro valor: '))
-valor2 = int(input('Digite o segundo valor: '))
-valor3 = int(input('Digite o terceiro valor: '))
+valor1 = float(input('Digite o primeiro valor: '))
+valor2 = float(input('Digite o segundo valor: '))
+valor3 = float(input('Digite o terceiro valor: '))
 
 print('Soma dos valores é: ', valor1 + valor2 + valor3)
 print('Média dos valores é: ', (valor1 + valor2 + valor3)/3)
